+++ common/include/globals.h
@@ -43,7 +43,7 @@
 #define ARK_BIN_MAX_SIZE 0x8000
 #define ARK_MAJOR_VERSION 4
 #define ARK_MINOR_VERSION 9
-#define ARK_MICRO_VERSION 2
+#define ARK_MICRO_VERSION 3
 
 /*
 First two bits identify the device (PSP or PS Vita)