+++ core/systemctrl/src/loadercore.c
@@ -35,6 +35,7 @@
 #include "cryptography.h"
 #include "rebootconfig.h"
 #include "graphics.h"
+#include "libs/colordebugger/colordebugger.h"
 
 // init.prx Text Address
 unsigned int sceInitTextAddr = 0;
@@ -45,6 +46,10 @@ int pluginLoaded = 0;
 // Real Executable Check Function Pointer
 int (* ProbeExec1)(u8 *buffer, int *check) = NULL;
 int (* ProbeExec2)(u8 *buffer, int *check) = NULL;
+
+// Sony PRX Decrypter Function Pointer
+int (* SonyPRXDecrypt)(void *, unsigned int, unsigned int *) = NULL;
+int (* origCheckExecFile)(unsigned char * addr, void * arg2) = NULL;
 
 // init.prx Custom sceKernelStartModule Handler
 int (* customStartModule)(int modid, SceSize argsize, void * argp, int * modstatus, SceKernelSMOption * opt) = NULL;
@@ -147,6 +152,31 @@ int _ProbeExec2(u8 *buffer, int *check)
     return result;
 }
 
+// PRO GZIP Decrypt Support
+int PROPRXDecrypt(void * prx, unsigned int size, unsigned int * newsize)
+{
+
+    // GZIP Packed PRX File
+    if ( (_lb((unsigned)prx + 0x150) == 0x1F && _lb((unsigned)prx + 0x151) == 0x8B)
+            || (*(unsigned int *)(prx + 0x130) == 0xC01DB15D) )
+    {
+        // Read GZIP Size
+        unsigned int compsize = *(unsigned int *)(prx + 0xB0);
+        
+        // Return GZIP Size
+        *newsize = compsize;
+        
+        // Remove PRX Header
+        memcpy(prx, prx + 0x150, compsize);
+        
+        // Fake Decrypt Success
+        return 0;
+    }
+    
+    // Decrypt Sony PRX Files
+    return SonyPRXDecrypt(prx, size, newsize);
+}
+
 // Executable File Check
 int KernelCheckExecFile(unsigned char * buffer, int * check)
 {
@@ -214,6 +244,7 @@ int InitKernelStartModule(int modid, SceSize argsize, void * argp, int * modstat
 int patch_sceKernelStartModule_in_bootstart(int (* bootstart)(SceSize, void *), void * argp)
 {
     
+    /*
     u32 StartModule = JUMP(FindFunction("sceModuleManager", "ModuleMgrForUser", 0x50F0C1EC));
 
     u32 addr = (u32)bootstart;
@@ -226,21 +257,35 @@ int patch_sceKernelStartModule_in_bootstart(int (* bootstart)(SceSize, void *), 
             patches--;
         }
     }
+    */
+    
+        /*
+		.sceInitBootStartCall = ,
+		.sceKernelLinkLibraryEntries = 0x00001110,
+		.sceKernelLinkLibraryEntriesForUser = 0x000025A4,
+		.sceKernelIcacheClearAll = 0x0000748C,
+		*/
+    
+    u32 addr = ((u32)bootstart) + 0x00001C3C - 0x00001A4C;
+	_sw(JUMP(InitKernelStartModule), addr);
+    _sw(NOP, addr + 4);
     
     // Passthrough
     return bootstart(4, argp);
 }
 
-// Sony PRX Decrypter Function Pointer
-int (* SonyPRXDecrypt)(void *, unsigned int, unsigned int *) = NULL;
-int (* origCheckExecFile)(unsigned char * addr, void * arg2) = NULL;
+static void myDebug(int a0){
+    if (a0<0) doBreakpoint();
+    return;
+}
+
 // Patch Loader Core Module
 SceModule2* patchLoaderCore(void)
 {
 
     // Find Module
     SceModule2* mod = (SceModule2 *)sceKernelFindModuleByName("sceLoaderCore");
-    
+
     // Fetch Text Address
     u32 start_addr = mod->text_addr;
     u32 topaddr = mod->text_addr+mod->text_size;
@@ -318,10 +363,9 @@ SceModule2* patchLoaderCore(void)
         else if (_lw(addr) == JAL(ProbeExec2))
             _sw (JAL(_ProbeExec2), addr);
     }
-    
     // Flush Cache
     flushCache();
-    
+
     return mod;
 }