+++ loader/rebootex/pspbtcnf.c
@@ -227,6 +227,9 @@ int UnpackBootConfigPatched(char **p_buffer, int length)
         if (IS_PSP(ark_config)){
             newsize = patch_bootconf_psp(buffer, length);
             if (newsize > 0) result = newsize;
+            // Insert Stargate No-DRM Engine
+            newsize = AddPRX(buffer, "/kd/me_wrapper.prx", PATH_STARGATE+sizeof(PATH_FLASH0)-2, GAME_RUNLEVEL | UMDEMU_RUNLEVEL);
+            if (newsize > 0) result = newsize;
         }
         else if (IS_VITA(ark_config)){
             if (IS_VITA_POPS(ark_config)){
@@ -236,19 +239,12 @@ int UnpackBootConfigPatched(char **p_buffer, int length)
             else{
                 newsize = patch_bootconf_vita(buffer, length);
                 if (newsize > 0) result = newsize;
+                // Insert Stargate No-DRM Engine
+                newsize = AddPRX(buffer, "/kd/kermit_me_wrapper.prx", PATH_STARGATE+sizeof(PATH_FLASH0)-2, GAME_RUNLEVEL | UMDEMU_RUNLEVEL);
+                if (newsize > 0) result = newsize;
             }
         }
         else colorDebug(0xff); // unknown device (?), don't touch it
-    }
-    
-    // Insert Stargate No-DRM Engine
-    if (SearchPrx(buffer, "/vsh/me_wrapper.prx") >= 0){
-        newsize = AddPRX(buffer, "/kd/me_wrapper.prx", PATH_STARGATE+sizeof(PATH_FLASH0)-2, GAME_RUNLEVEL | UMDEMU_RUNLEVEL);
-        if (newsize > 0) result = newsize;
-    }
-    else if (SearchPrx(buffer, "/vsh/kermit_me_wrapper.prx") >= 0){
-        newsize = AddPRX(buffer, "/kd/kermit_me_wrapper.prx", PATH_STARGATE+sizeof(PATH_FLASH0)-2, GAME_RUNLEVEL | UMDEMU_RUNLEVEL);
-        if (newsize > 0) result = newsize;
     }
     
     // Insert VSHControl